cli: move report rendering out of binary
