+++ apps/cli/cmd/newsjack/detector_command.go
@@ -41,7 +41,6 @@ type detectorOptions struct {
 	MonitorName                     string
 	IncludeAllScored                bool
 	ScoredOutput                    string
-	Emit                            string
 }
 
 func cmdDetector(args []string, stdout, stderr io.Writer) int {
@@ -107,7 +106,6 @@ func parseDetectorRun(args []string, stderr io.Writer) (detectorOptions, int) {
 		MinQueuePriority:                defaultMinQueuePriority,
 		MinMajorNews:                    defaultMinMajorNews,
 		Limit:                           20,
-		Emit:                            "json",
 	}
 	fs.Var(&topics, "topic", "Topic to monitor. Repeatable.")
 	fs.StringVar(&opts.ProfilePath, "profile", "", "Path to monitor profile JSON")
@@ -140,12 +138,11 @@ func parseDetectorRun(args []string, stderr io.Writer) (detectorOptions, int) {
 	fs.StringVar(&opts.MonitorName, "monitor-name", "", "Monitor name")
 	fs.BoolVar(&opts.IncludeAllScored, "include-all-scored", false, "Include all scored signals in debug output")
 	fs.StringVar(&opts.ScoredOutput, "scored-output", "", "Write all scored signals to a separate JSON artifact")
-	fs.StringVar(&opts.Emit, "emit", "json", "json or brief")
 	valueFlags := stringSet([]string{
 		"topic", "profile", "sources", "feed-url", "feed-file", "depth", "lookback-days", "max-age-hours",
 		"x-news-min-profile-match", "x-posts-min-profile-match", "profile-relevance-min-profile-match",
 		"major-news-min-profile-match", "x-trends-min-profile-match", "min-queue-priority", "min-major-news",
-		"lane-caps", "limit", "store", "monitor-name", "scored-output", "emit",
+		"lane-caps", "limit", "store", "monitor-name", "scored-output",
 	})
 	if err := fs.Parse(reorderIntermixedFlags(args, valueFlags)); err != nil {
 		return opts, 2
@@ -156,10 +153,6 @@ func parseDetectorRun(args []string, stderr io.Writer) (detectorOptions, int) {
 	opts.Query = fs.Args()
 	if opts.Depth != "quick" && opts.Depth != "default" && opts.Depth != "deep" {
 		fmt.Fprintln(stderr, "invalid --depth")
-		return opts, 2
-	}
-	if opts.Emit != "json" && opts.Emit != "brief" {
-		fmt.Fprintln(stderr, "invalid --emit")
 		return opts, 2
 	}
 	return opts, 0
