--- fixtures/golden/go-port/summarize/run.md
@@ -1,42 +1,0 @@
-# Newsjack Newsjack Brief
-
-| key | value |
-|---|---|
-| Status | Detector preview only |
-| Generated | 2026-05-25 21:25 UTC |
-| Detector candidates | 2 |
-
-## Candidate Preview
-
-1. **Regulators open inquiry tied to AI customer support**
-   - Why surfaced: profile relevance; queue 70.8, profile 0.4, major 0.
-   - news search: [Regulators open inquiry tied to AI customer support](https://example.com/news/60f5725c) (2026-05-25)
-2. **Experts are reacting to AI customer support**
-   - Why surfaced: x posts; queue 64, profile 0.4, major 0.
-   - x: [Experts are reacting to AI customer support](https://x.com/example/status/f7f584be) (2026-05-25)
-
-## What Was Scanned
-
-- **Profile:** (unknown)
-- **Queries:** AI customer support
-
-## Appendix: Provenance
-
-### Pipeline
-
-| key | value |
-|---|---|
-| detector | pending - candidates.json |
-| coarse_relevance | pending - coarse_relevance_decisions.json |
-| relevance_apply | pending - relevant_candidates.json |
-| story_origin | pending - origin_findings.json |
-| freshness_gate | pending - targeted_candidates.json |
-| final_report | pending - final_report.md |
-
-### Detector Counts
-
-| key | value |
-|---|---|
-| scored | 2 |
-| selected | 2 |
-| source_errors | 0 |
