+++ fixtures/newsjack-detector-agent/scripts/run-detector-json.sh
@@ -25,5 +25,4 @@ exec "$SCRIPT_DIR/with-fixture-env.sh" \
   --sources news_search,x \
   --lookback-days 1 \
   --depth quick \
-  --save \
-  --emit json
+  --save
