+++ fixtures/newsjack-detector-agent/scripts/run-mock-detector.sh
@@ -22,5 +22,4 @@ exec "$SCRIPT_DIR/with-fixture-env.sh" \
   "$QUERY" \
   --profile "$PROFILE" \
   "$@" \
-  --mock \
-  --emit json
+  --mock
